+++ search/src/mapwisefox/search/adapters/query_builder/_science_direct.py
@@ -1,4 +1,4 @@
-from mapwisefox.search import (
+from mapwisefox.search.query_builder import (
     EvidenceAttributes,
     EvidenceTypes,
     YearRangeExpr,
