+++ search/src/mapwisefox/search/adapters/query_builder/_wos.py
@@ -1,4 +1,4 @@
-from mapwisefox.search import (
+from mapwisefox.search.query_builder import (
     EvidenceAttributes,
     EvidenceTypeExpr,
     EvidenceTypes,
