+++ search/src/mapwisefox/search/query_builder/__init__.py
@@ -1,0 +1,47 @@
+from ._enum import EvidenceAttributes, EvidenceTypes, SubjectAreas
+from ._expr import (
+    Expr,
+    StrExpr,
+    AttrExpr,
+    TitleExpr,
+    TitleAbsExpr,
+    TitleAbsKeysExpr,
+    AbstractExpr,
+    SubjectAreaExpr,
+    SubjectAreasExpr,
+    EvidenceTypeExpr,
+    EvidenceTypesExpr,
+    LanguageExpr,
+    LanguagesExpr,
+    LogicalOperators,
+    YearRangeExpr,
+    AuthorKeysExpr,
+    MultipleKeysExpr,
+    NaryExpr,
+)
+from ._qb import QueryBuilder
+
+__all__ = [
+    "QueryBuilder",
+    "Expr",
+    "StrExpr",
+    "AttrExpr",
+    "TitleExpr",
+    "TitleAbsExpr",
+    "TitleAbsKeysExpr",
+    "AbstractExpr",
+    "SubjectAreaExpr",
+    "SubjectAreasExpr",
+    "EvidenceTypeExpr",
+    "EvidenceTypesExpr",
+    "LanguageExpr",
+    "LanguagesExpr",
+    "LogicalOperators",
+    "YearRangeExpr",
+    "AuthorKeysExpr",
+    "MultipleKeysExpr",
+    "NaryExpr",
+    "EvidenceAttributes",
+    "EvidenceTypes",
+    "SubjectAreas",
+]
