+++ search/src/mapwisefox/search/query_builder/_enum.py
